refactor(settings): remove orphan sandbox query module
